fix(others): Update database
